ionq dict type check-in
