+++ .gitignore
@@ -135,4 +135,5 @@ dmypy.json
 # Testing
 p0.prof
 None/
+**/Untitled*.ipynb
