+++ qbraid/transpiler/conversions/qasm2/__init__.py
@@ -23,11 +23,13 @@ Functions
     qasm2_to_pytket
     qasm2_to_qiskit
     qasm2_to_qasm3
+    qasm2_to_ionq
 
 """
 from .qasm2_to_cirq import qasm2_to_cirq
+from .qasm2_to_ionq import qasm2_to_ionq
 from .qasm2_to_pytket import qasm2_to_pytket
 from .qasm2_to_qasm3 import qasm2_to_qasm3
 from .qasm2_to_qiskit import qasm2_to_qiskit
 
-__all__ = ["qasm2_to_cirq", "qasm2_to_pytket", "qasm2_to_qasm3", "qasm2_to_qiskit"]
+__all__ = ["qasm2_to_cirq", "qasm2_to_pytket", "qasm2_to_qasm3", "qasm2_to_qiskit", "qasm2_to_ionq"]
