+++ qbraid/transpiler/conversions/qasm2/qasm2_to_cirq.py
@@ -12,6 +12,8 @@
 Module for conversions from QASM 2 to Cirq Circuits
 
 """
+from __future__ import annotations
+
 from typing import TYPE_CHECKING
 
 from qbraid_core._import import LazyLoader
@@ -30,7 +32,7 @@ if TYPE_CHECKING:
 
 
 @weight(1)
-def qasm2_to_cirq(qasm: str) -> "cirq.Circuit":
+def qasm2_to_cirq(qasm: str) -> cirq.Circuit:
     """Returns a Cirq circuit equivalent to the input QASM string.
 
     Args:
