+++ qbraid/transpiler/conversions/qasm2/qasm2_to_pytket.py
@@ -12,6 +12,8 @@
 Module containing functions to convert between OpenQASM 2 and PyTKET.
 
 """
+from __future__ import annotations
+
 from typing import TYPE_CHECKING
 
 from qbraid_core._import import LazyLoader
@@ -25,7 +27,7 @@ if TYPE_CHECKING:
 
 
 @weight(1)
-def qasm2_to_pytket(qasm: str) -> "pytket.circuit.Circuit":
+def qasm2_to_pytket(qasm: str) -> pytket.circuit.Circuit:
     """Returns a pytket circuit equivalent to the input OpenQASM 2 string.
 
     Args:
