+++ qbraid/transpiler/conversions/qasm2/qasm2_to_qasm3.py
@@ -19,6 +19,7 @@ from qbraid._version import __version__ as qbraid_version
 from qbraid.passes.qasm2.decompose import _decompose_rxx_instr
 from qbraid.passes.qasm3.format import remove_unused_gates
 from qbraid.programs import get_qasm_type_alias
+from qbraid.programs.typer import Qasm2String, Qasm3String
 from qbraid.transpiler.annotations import weight
 
 
@@ -100,7 +101,7 @@ def _convert_line_to_qasm3(line: str) -> str:
 
 
 @weight(0.7)
-def qasm2_to_qasm3(qasm_str: str) -> str:
+def qasm2_to_qasm3(qasm_str: Qasm2String) -> Qasm3String:
     """Convert a OpenQASM 2.0 string to OpenQASM 3.0 string
 
     Args:
