+++ qbraid/transpiler/conversions/qasm2/qasm2_to_qiskit.py
@@ -12,6 +12,8 @@
 Module defining Qiskit OpenQASM conversions
 
 """
+from __future__ import annotations
+
 from typing import TYPE_CHECKING
 
 from qbraid_core._import import LazyLoader
@@ -25,7 +27,7 @@ if TYPE_CHECKING:
 
 
 @weight(1)
-def qasm2_to_qiskit(qasm: str) -> "qiskit_.QuantumCircuit":
+def qasm2_to_qiskit(qasm: str) -> qiskit_.QuantumCircuit:
     """Returns a Qiskit circuit equivalent to the input OpenQASM 2 string.
 
     Args:
